HSEARCH-3200 Extract common min should match options into reused interface
